backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic[email]==2.9.1
python-multipart==0.0.9
anthropic==0.40.0
httpx==0.27.2
pdfplumber==0.11.4
python-docx==1.1.2
aiofiles==24.1.0
python-dotenv==1.0.1
